move LineFormatter, add basic OnTypeFormatting tests, add LineFormatter tests from vscode-python
